+++ inc/fit-db.php
@@ -100,6 +100,10 @@ function get_unique($fit) {
 			$uniquep['implants'][(int)$i['typeid']] = true;
 		}
 
+		foreach($preset['beacons'] as $b) {
+			$uniquep['beacons'][(int)$b['typeid']] = true;
+		}
+
 		$unique['presets'][] = $uniquep;
 	}
 
@@ -445,6 +449,21 @@ function commit_fitting(&$fit, &$error = null) {
 					$fittinghash,
 					$presetid,
 					$i['typeid'],
+				)
+			);
+
+			if($ret === false) {
+				return false;
+			}
+		}
+
+		foreach($preset['beacons'] as $typeid => $b) {
+			$ret = \Osmium\Db\query_params(
+				'INSERT INTO osmium.fittingbeacons (fittinghash, presetid, typeid) VALUES ($1, $2, $3)',
+				array(
+					$fittinghash,
+					$presetid,
+					$b['typeid'],
 				)
 			);
 
@@ -1036,6 +1055,16 @@ function get_fitting($fittinghash) {
 		);
 		while($implant = \Osmium\Db\fetch_row($implantsq)) {
 			add_implant($fit, $implant[0]);
+		}
+
+		$beaconsq = \Osmium\Db\query_params(
+			'SELECT typeid
+			FROM osmium.fittingbeacons
+			WHERE fittinghash = $1 AND presetid = $2',
+			array($fittinghash, $preset['presetid'])
+		);
+		while($beacon = \Osmium\Db\fetch_row($beaconsq)) {
+			add_beacon($fit, $beacon[0]);
 		}
 	}